[FEAT] 20221005 오후 수업

+++ django2/articles/views.py
@@ -27,7 +27,9 @@ def create(request):
         # create
         form = ArticleForm(request.POST)
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
             return redirect("articles:detail", article.pk)
     else:
         # new
@@ -52,10 +54,12 @@ def detail(request, pk):
 
 @require_POST
 def delete(request, pk):
+    article = Article.objects.get(pk=pk)
     if request.user.is_authenticated:
-        article = Article.objects.get(pk=pk)
-        article.delete()
-    return redirect("articles:index")
+        if article.user == request.user:
+            article.delete()
+            return redirect("articles:index")
+    return redirect("articles:detail", article.pk)
 
 # def edit(request, pk):
 #     article = Article.objects.get(pk=pk)
@@ -70,26 +74,42 @@ def delete(request, pk):
 @require_http_methods(["GET", "POST"])
 def update(request, pk):
     article = Article.objects.get(pk=pk)
-    if request.method == "POST":
-        # update
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid:
-            form.save()
-            return redirect("articles:detail", article.pk)
+    if article.user == request.user:
+        if request.method == "POST":
+            # update
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid:
+                form.save()
+                return redirect("articles:detail", article.pk)
+        else:
+            # edit
+            form = ArticleForm(instance=article)
     else:
-        # edit
-        form = ArticleForm(instance=article)
+        return redirect("articles:index")
     context = {
         "article" : article,
         "form" : form,
     }
     return render(request, "articles/update.html", context)
 
+@require_POST
 def comments_create(request, pk):
-    article =Article.objects.get(pk=pk)
-    comment_form = CommentForm(request.POST)
-    if comment_form.is_valid():
-        comment = comment_form.save(commit=False)
-        comment.article = article
-        comment.save()
-    return redirect("articles:detail", article.pk)
+    if request.user.is_authenticated:
+        article =Article.objects.get(pk=pk)
+        comment_form = CommentForm(request.POST)
+        if comment_form.is_valid():
+            comment = comment_form.save(commit=False)
+            comment.article = article
+            comment.user = request.user
+            comment.save()
+        return redirect("articles:detail", article.pk)
+    return redirect("accounts:login")
+
+@require_POST
+def comments_delete(request, article_pk, comment_pk):
+    if request.user.is_authenticated:
+        comment = Comment.objects.get(pk=comment_pk)
+        if comment.user == request.user:
+            comment.delete()
+        return redirect("articles:detail", article_pk)
+    return redirect("articles:login")
